Pull `prime_fmod_size` policy class into its own dedicated header
